validation left

+++ php/deleteProducts.php
@@ -12,19 +12,21 @@ function massDelete($server = 'localhost', $user = 'id19542860_root')
     $postData = json_decode(file_get_contents("php://input"), true);
     $skuArray = $postData['selected'] ?? [];
     try {
+        if (!count($skuArray))
+            throw new PDOException(' nothing selected!');
         $conn = new \PDO("mysql:host=$servername;dbname=id19542860_testdb", $username, $password);
         $conn->setAttribute(\PDO::ATTR_ERRMODE, \PDO::ERRMODE_EXCEPTION);
+
         foreach ($skuArray as $sku) {
             $sql = "DELETE FROM products WHERE sku = '$sku'";
             $result = $conn->prepare($sql);
             $result->execute();
-            // echo "Success!";
         }
-        // echo "Success!";
     } catch (\PDOException $e) {
-        echo "Connection failed: " . $e->getMessage();
+        $errorMessage = $e->getMessage();
+        $response = array('error' => "Failed to delete: " . $errorMessage);
+        echo json_encode($response);
     }
-    // header("Location: /");
 }
 
 massDelete('localhost', 'id19542860_root', $_POST);
